agent: move NotifyGroup into the agent pkg
