+++ agent/notify_test.go
@@ -1,4 +1,4 @@
-package state
+package agent
 
 import (
 	"testing"
